Document the stream tests, and split what that pushed over the cap
AGENTS.md requires NumPy-style docstrings on public interfaces, and the
test functions in these modules were only partly converted:
`test_rust_streams.py` documented its four module-level pump tests but
none of the seven `TestRustConsumeStream` methods, and `test_rust_errno.py`
documented its two private Win32 helpers but none of its five tests.
CodeRabbit named three functions; converting only those would have left
both modules exactly as inconsistent, in different places.

Document every test function in both, and keep the explanatory prose that
was already there — it is the more valuable half — by moving it under
`Notes`, below the sections a reader scans for first. The paragraph in
`test_propagates_io_errors` distinguishing its platform-neutral remit from
`test_rust_errno.py`'s stricter conversion contract is preserved verbatim.

That pushed both modules past the 400-line ceiling pylint enforces:
`test_rust_errno.py` to 413 and `test_rust_streams.py` to 414. Shaving the
`Returns` sections to fit would have made the documentation inconsistent
again in order to satisfy a rule about file size, so split along the seams
the modules had already drawn for themselves:

- `test_rust_errno.py` said in its own docstring that the conversion has a
  POSIX arm and a Windows arm that hand the number over differently. The
  Windows arm shares no fixture with the POSIX one, so it moves whole to
  `test_rust_errno_windows.py` with the `ctypes` machinery it needs.
- `test_rust_streams.py` covered two entry points and had drawn the line
  between them as a class. `TestRustConsumeStream` and its
  `_consume_payload` helper move to `test_rust_consume_stream.py`.

`INVALID_FD_ERRNOS` and its derived match pattern are now wanted by both
the pump and consume modules, so they move to `tests/helpers/stream_pipes`
rather than being restated. No test body changed. Modules land at 165, 268,
228 and 217 lines.

Cross-references updated in `test_rust_errno.py`,
`test_rust_streams_boundary_property.py`, the consume integration guard,
and two passages of the developers' guide. Wheel-manifest snapshot
regenerated for the two new files.
