+++ cuprum/unittests/test_rust_streams.py
@@ -1,7 +1,8 @@
-"""Unit tests for the Rust stream pump and consumer.
+"""Unit tests for the Rust stream pump.
 
-These tests validate the optional Rust-backed pump and consume behaviour and
-error handling.
+These tests validate the optional Rust-backed pump's transfer behaviour and
+error handling. The consume entry point is covered in
+``test_rust_consume_stream.py``.
 
 Example
 -------
@@ -13,58 +14,23 @@ from __future__ import annotations
 import contextlib
 import errno
 import os
-import re
 import typing as typ
 
 import pytest
 
 from tests.helpers.stream_pipes import (
+    INVALID_FD_ERRNOS,
+    INVALID_FD_MESSAGE_RE,
     _pipe_pair,
     _pump_rust_stream_payload,
     _safe_close,
     feed_source_pipe,
 )
 
-# An unusable descriptor reports `EBADF` on POSIX and `EINVAL` where Windows
-# rejects the handle instead.
-_INVALID_FD_ERRNOS = (errno.EBADF, errno.EINVAL)
-
-# `pytest.raises(OSError)` needs a `match` to satisfy ruff PT011, but
-# `strerror` comes from the C library and is translated under a non-English
-# locale. Anchor on the `[Errno N]` prefix CPython formats itself, which no
-# locale changes.
-_INVALID_FD_MESSAGE_RE = "|".join(
-    re.escape(f"[Errno {code}]") for code in _INVALID_FD_ERRNOS
-)
-
 if typ.TYPE_CHECKING:
     from types import ModuleType
 
 
-def _consume_payload(
-    streams: ModuleType,
-    payload: bytes,
-    **kwargs: object,
-) -> str:
-    """Consume payload through the Rust stream and return decoded output."""
-    with contextlib.ExitStack() as stack:
-        read_fd, write_fd = os.pipe()
-        stack.callback(_safe_close, read_fd)
-        stack.callback(_safe_close, write_fd)
-        view = memoryview(payload)
-        while view:
-            written = os.write(write_fd, view)
-            assert written > 0, "expected os.write to make progress"
-            view = view[written:]
-        _safe_close(write_fd)
-        forwarded_kwargs = dict(kwargs)
-        if forwarded_kwargs.get("buffer_size") is None:
-            forwarded_kwargs.pop("buffer_size", None)
-
-        return typ.cast(
-            "str",
-            streams.rust_consume_stream(read_fd, **forwarded_kwargs),
-        )
 @pytest.mark.parametrize(
     ("test_id", "payload", "buffer_size"),
     [
@@ -95,6 +61,10 @@ def test_rust_pump_stream_transfers_data(
         The payload to transfer through the pump.
     buffer_size : int | None
         Optional buffer size parameter; None uses default.
+
+    Returns
+    -------
+    None
     """
     output, transferred = _pump_rust_stream_payload(
         rust_streams, payload, buffer_size=buffer_size
@@ -115,6 +85,10 @@ def test_rust_pump_stream_raises_on_invalid_buffer(
     ----------
     rust_streams : ModuleType
         The Rust streams module fixture.
+
+    Returns
+    -------
+    None
     """
     with contextlib.ExitStack() as stack:
         in_read, in_write = os.pipe()
@@ -133,6 +107,10 @@ def test_rust_pump_stream_propagates_io_errors(
     ----------
     rust_streams : ModuleType
         The Rust streams module fixture.
+
+    Returns
+    -------
+    None
     """
     with contextlib.ExitStack() as stack:
         read_fd, write_fd = os.pipe()
@@ -141,10 +119,10 @@ def test_rust_pump_stream_propagates_io_errors(
         _safe_close(read_fd)
         with pytest.raises(
             OSError,
-            match=_INVALID_FD_MESSAGE_RE,
+            match=INVALID_FD_MESSAGE_RE,
         ) as excinfo:
             rust_streams.rust_pump_stream(read_fd, write_fd)
-    assert excinfo.value.errno in _INVALID_FD_ERRNOS, (
+    assert excinfo.value.errno in INVALID_FD_ERRNOS, (
         "expected errno to indicate an invalid file descriptor/handle, "
         f"found {excinfo.value.errno!r}"
     )
@@ -160,11 +138,9 @@ def test_rust_pump_stream_ignores_broken_pipe(
     rust_streams : ModuleType
         The Rust streams module fixture.
 
-    Raises
-    ------
-    OSError
-        If the pump fails for a reason other than a broken pipe or a
-        connection reset.
+    Returns
+    -------
+    None
     """
     payload = b"x" * (64 * 1024)
     with _pipe_pair() as (in_read, in_write, out_read, out_write):
@@ -187,146 +163,3 @@ def test_rust_pump_stream_ignores_broken_pipe(
 
     assert isinstance(transferred, int), "expected transfer count to be integer"
     assert transferred <= len(payload), "expected transfer count to be bounded"
-
-
-class TestRustConsumeStream:
-    """Coverage for Rust-backed consume stream helpers."""
-
-    @staticmethod
-    def _consume(
-        rust_streams: ModuleType,
-        payload: bytes,
-        **kwargs: object,
-    ) -> str:
-        """Consume payload via the Rust helper."""
-        return _consume_payload(rust_streams, payload, **kwargs)
-
-    @pytest.mark.parametrize(
-        ("test_id", "payload", "buffer_size"),
-        [
-            ("ascii_explicit_default", b"rust-consume-stream", 65536),
-            ("multibyte_split", b"snowman \xe2\x98\x83", 2),
-        ],
-        ids=["ascii_explicit_default", "multibyte_split"],
-    )
-    def test_decodes_payload(
-        self,
-        rust_streams: ModuleType,
-        test_id: str,
-        payload: bytes,
-        buffer_size: int,
-    ) -> None:
-        """Validate rust_consume_stream decodes UTF-8 payloads."""
-        output = self._consume(rust_streams, payload, buffer_size=buffer_size)
-        expected = payload.decode("utf-8", errors="replace")
-        assert output == expected, (
-            f"expected decoded output to match Python replace semantics ({test_id})"
-        )
-
-    def test_uses_default_buffer_size(
-        self,
-        rust_streams: ModuleType,
-    ) -> None:
-        """Validate rust_consume_stream uses the default buffer size."""
-        payload = b"rust-consume-default"
-        output = self._consume(rust_streams, payload)
-        expected = payload.decode("utf-8", errors="replace")
-        assert output == expected, "expected default buffer size to decode payload"
-
-    def test_replaces_invalid_bytes(
-        self,
-        rust_streams: ModuleType,
-    ) -> None:
-        """Ensure rust_consume_stream replaces invalid UTF-8 bytes."""
-        payload = b"valid-\xff\xfe-end"
-        output = self._consume(rust_streams, payload, buffer_size=3)
-        expected = payload.decode("utf-8", errors="replace")
-        assert output == expected, "expected invalid bytes to be replaced"
-
-    def test_replaces_incomplete_sequence(
-        self,
-        rust_streams: ModuleType,
-    ) -> None:
-        """Ensure rust_consume_stream replaces incomplete UTF-8 sequences."""
-        payload = b"trail-\xe2\x98"
-        output = self._consume(rust_streams, payload, buffer_size=2)
-        expected = payload.decode("utf-8", errors="replace")
-        assert output == expected, "expected incomplete sequence to be replaced"
-
-    @staticmethod
-    def test_does_not_close_fd(
-        rust_streams: ModuleType,
-    ) -> None:
-        """Ensure rust_consume_stream does not close the underlying FD.
-
-        Raises
-        ------
-        OSError
-            If reading the file descriptor fails for a reason other than
-            it being closed.
-        """
-        with contextlib.ExitStack() as stack:
-            read_fd, write_fd = os.pipe()
-            stack.callback(_safe_close, read_fd)
-            stack.callback(_safe_close, write_fd)
-            os.write(write_fd, b"non-destructive")
-            _safe_close(write_fd)
-            output = rust_streams.rust_consume_stream(read_fd)
-            assert output == "non-destructive"
-
-            try:
-                os.read(read_fd, 0)
-            except OSError as exc:
-                if exc.errno == errno.EBADF:
-                    pytest.fail(
-                        "rust_consume_stream must not close the file descriptor"
-                    )
-                raise
-
-    @staticmethod
-    def test_rejects_invalid_buffer(
-        rust_streams: ModuleType,
-    ) -> None:
-        """Verify rust_consume_stream rejects invalid buffer sizes."""
-        with contextlib.ExitStack() as stack:
-            read_fd, write_fd = os.pipe()
-            stack.callback(_safe_close, read_fd)
-            stack.callback(_safe_close, write_fd)
-            _safe_close(write_fd)
-            with pytest.raises(ValueError, match="buffer_size"):
-                rust_streams.rust_consume_stream(read_fd, buffer_size=0)
-
-    @staticmethod
-    def test_propagates_io_errors(
-        rust_streams: ModuleType,
-    ) -> None:
-        """Verify rust_consume_stream raises OSError on I/O failure.
-
-        The consume-side counterpart to
-        ``test_rust_pump_stream_propagates_io_errors``: a read that cannot be
-        performed surfaces as an ``OSError`` whose ``errno`` names an unusable
-        descriptor. Either ``EBADF`` or ``EINVAL`` is accepted, because Windows
-        reports an invalid handle rather than a bad POSIX descriptor, so the
-        assertion holds wherever the entry points build.
-
-        The Rust-side tests over ``consume_stream_files`` exercise the
-        read-and-decode loop below the PyO3 boundary and so cannot observe this
-        translation at all. ``test_rust_errno.py`` pins the stricter conversion
-        contract — the exact POSIX number, ``strerror``, and subclass selection
-        — and skips on Windows for that reason; this test stays platform
-        neutral and lives beside the other consume behaviour.
-        """
-        with contextlib.ExitStack() as stack:
-            read_fd, write_fd = os.pipe()
-            stack.callback(_safe_close, read_fd)
-            stack.callback(_safe_close, write_fd)
-            _safe_close(read_fd)
-            with pytest.raises(
-                OSError,
-                match=_INVALID_FD_MESSAGE_RE,
-            ) as excinfo:
-                rust_streams.rust_consume_stream(read_fd)
-        assert excinfo.value.errno in _INVALID_FD_ERRNOS, (
-            "expected errno to indicate an invalid file descriptor/handle, "
-            f"found {excinfo.value.errno!r}"
-        )
